requirements.txt
# Core dependencies
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0

# Exchange connectivity
ccxt==4.2.25

# Technical analysis
ta-lib==0.4.28
pandas-ta==0.3.14b

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
redis==5.0.1
influxdb-client==1.40.0

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.3.0

# Async and networking
aiohttp==3.9.1
websockets==12.0
asyncio==3.4.3

# Monitoring and logging
prometheus-client==0.19.0
python-telegram-bot==21.0.1

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0

# Telegram bot
python-telegram-bot==20.7

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0